chore(tsconfig): add "*.js" to include, to fix eslint "ProjectService" error
added it only to the "dev" tsconfig.
The tsconfig.build.json, which is used for build output does not require it.

eslint was previosuly reporting on itself:
"TriliumNextNotes/eslint.config.js was not found by the project service. Consider either including it in the tsconfig.json or including it in allowDefaultProject"