chore(nx/forge): add strange workaround to solve node modules
Here we need to ensure that there is no `node_modules` in
`apps/desktop` since electron-forge will pick it up and use only that,
meaning that all the hoisted dependencies available in the project root
will not be accessible and it will error out.

See https://github.com/pnpm/pnpm/issues/7880 for more information.